refactor: consolidate workflow/ into workflows/ (#249)
Move checkpoint.py from assemblyzero/workflow/ (singular) to
assemblyzero/workflows/ (plural) to eliminate the confusing
duplicate directory. Update all imports and remove the old package
from the circular import test list.

Closes #249
