--- assemblyzero/workflow/__init__.py
@@ -1,8 +1,0 @@
-"""AssemblyZero workflow utilities.
-
-This package contains workflow-related utilities including checkpoint management.
-"""
-
-from assemblyzero.workflow.checkpoint import get_checkpoint_db_path, get_repo_root
-
-__all__ = ["get_checkpoint_db_path", "get_repo_root"]
