+++ assemblyzero/workflows/__init__.py
@@ -1,1 +1,5 @@
 """AssemblyZero workflow modules."""
+
+from assemblyzero.workflows.checkpoint import get_checkpoint_db_path, get_repo_root
+
+__all__ = ["get_checkpoint_db_path", "get_repo_root"]
